Atualização Automática via bash shell

+++ criar_box.sh
@@ -1,34 +1,38 @@
 #!/bin/bash
 
-echo "==> Atualizar os repositórios..."
-sudo apt update
+if ! command -v vboxmanage &> /dev/null
+then
+    echo "==> Atualizar os repositórios..."
+    sudo apt update
 
-echo "==> Instalar o VirtualBox"
-sudo apt install -y virtualbox
-sudo apt install -y virtualbox-guest-dkms virtualbox-guest-x11
-sudo apt install -y virtualbox-guest-additions-iso
+    echo "==> Instalar o VirtualBox"
+    sudo apt install -y virtualbox
+    sudo apt install -y virtualbox-guest-dkms virtualbox-guest-x11
+    sudo apt install -y virtualbox-guest-additions-iso
 
-echo "==> Instalar o Extension Pack do VirtualBox"
-wget https://download.virtualbox.org/virtualbox/6.1.18/Oracle_VM_VirtualBox_Extension_Pack-6.1.18.vbox-extpack
-sudo vboxmanage extpack install Oracle_VM_VirtualBox_Extension_Pack-6.1.18.vbox-extpack --accept-license=33d7284dc4a0ece381196fda3cfe2ed0e1e8e7ed7f27b9a9ebc4ee22e24bd23c
-rm Oracle_VM_VirtualBox_Extension_Pack-6.1.18.vbox-extpack
+    echo "==> Instalar o Extension Pack do VirtualBox"
+    wget https://download.virtualbox.org/virtualbox/6.1.18/Oracle_VM_VirtualBox_Extension_Pack-6.1.18.vbox-extpack
+    sudo vboxmanage extpack install Oracle_VM_VirtualBox_Extension_Pack-6.1.18.vbox-extpack --accept-license=33d7284dc4a0ece381196fda3cfe2ed0e1e8e7ed7f27b9a9ebc4ee22e24bd23c
+    rm Oracle_VM_VirtualBox_Extension_Pack-6.1.18.vbox-extpack
 
-echo "==> Download Vagrant & Instalar"
-wget -nv https://releases.hashicorp.com/vagrant/2.2.9/vagrant_2.2.9_x86_64.deb
-sudo dpkg -i vagrant_2.2.9_x86_64.deb
-rm vagrant_2.2.9_x86_64.deb
+    echo "==> Download Vagrant & Instalar"
+    wget -nv https://releases.hashicorp.com/vagrant/2.2.9/vagrant_2.2.9_x86_64.deb
+    sudo dpkg -i vagrant_2.2.9_x86_64.deb
+    rm vagrant_2.2.9_x86_64.deb
 
-echo "==> Instalar plugins do Vagrant"
-vagrant plugin install vagrant-libvirt
-vagrant plugin install vagrant-disksize # Só funciona no Virtualbox
-vagrant plugin install vagrant-mutate
+    echo "==> Instalar plugins do Vagrant"
+    vagrant plugin install vagrant-libvirt
+    vagrant plugin install vagrant-disksize # Só funciona no Virtualbox
+    vagrant plugin install vagrant-mutate
 
-echo "==> Removendo pacotes do Ubuntu desnecessários"
-apt autoremove -y
+    echo "==> Removendo pacotes do Ubuntu desnecessários"
+    apt autoremove -y
+fi
 
 echo "==> Iniciando a box..."
-VAGRANT_VAGRANTFILE=Vagrantfile_Virtualbox vagrant up
-vagrant ssh <<EOF
+iniciar_box(){
+    VAGRANT_VAGRANTFILE=Vagrantfile_Virtualbox vagrant up
+    vagrant ssh <<EOF
 #!/bin/bash
 
 cd /vagrant
@@ -36,3 +40,20 @@ make iso
 cd ..
 
 EOF
+}
+
+if vagrant status | grep "not created" > /dev/null; then
+    iniciar_box
+elif vagrant status | grep "is running" > /dev/null; then
+    echo "[DEBUG] O VPS_DEV existe e está ligado. Destruir e começar de novo?"
+    vagrant destroy
+    iniciar_box
+elif vagrant status | grep "poweroff" > /dev/null; then
+    echo "[DEBUG] O VPS_DEV existe mas está desligado. Destruir e começar de novo..."
+    vagrant destroy -f
+    iniciar_box
+else
+    echo "[DEBUG] O VPS_DEV existe mas está com um status diferente..."
+    vagrant status
+    sleep 5
+fi
